backend/pyproject.toml
[project]
name = "wine_app_backend"
version = "0.1.0"
description = "Backend for Wine App"
readme = "README.md"
requires-python = ">=3.12"
license = "MIT"
authors = [
    { name = "Jun Liu", email = "example@example.com" },
]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.1",
    "pydantic>=2.6.1",
    "httpx>=0.26.0",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.27",
    "starlette>=0.36.3",
    "loguru>=0.7.2",
    "supabase>=2.3.0",
    "tenacity>=8.2.3",
    "pytest>=8.0.2",
    "pytest-asyncio>=0.23.5",
    "requests>=2.31.0",
    "types-requests>=2.31.0.20240218",
    "fake-useragent>=1.4.0", # For generating random user agents
    "lxml>=5.1.0", # For HTML parsing
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.2",
    "pytest-asyncio>=0.23.5",
    "ruff>=0.2.1",
    "black>=24.1.1",
    "ipykernel>=6.29.2",
    "isort>=5.13.2",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py312"
line-length = 88
select = ["E", "F", "I", "W"]
ignore = []
fixable = ["ALL"]
unfixable = []
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
per-file-ignores = {}

[tool.ruff.mccabe]
max-complexity = 10

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true

[tool.black]
line-length = 88
target-version = ["py312"]
include = '\.pyi?$'

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
python_classes = "Test*"
addopts = "-v --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests that require infrastructure (deselect with '-m \"not integration\"')",
]
